+++ packages/react-three-cannon-examples/src/demos/demo-Paused.tsx
@@ -3,13 +3,13 @@ import { Debug, Physics, useBox, usePlane } from '@react-three/cannon'
 import { Box, OrbitControls, Plane } from '@react-three/drei'
 import type { MeshStandardMaterialProps } from '@react-three/fiber'
 import { Canvas } from '@react-three/fiber'
-import type { FC } from 'react'
-import { useState } from 'react'
+import { useRef, useState } from 'react'
+import type { Mesh } from 'three'
 
 type GroundProps = Pick<MeshStandardMaterialProps, 'color'> & PlaneProps
 
-const Ground: FC<GroundProps> = ({ color, ...props }) => {
-  const [ref] = usePlane(() => ({ ...props }))
+function Ground({ color, ...props }: GroundProps): JSX.Element {
+  const [ref] = usePlane(() => ({ ...props }), useRef<Mesh>(null))
 
   return (
     <Plane args={[1000, 1000]} ref={ref}>
@@ -18,8 +18,8 @@ const Ground: FC<GroundProps> = ({ color, ...props }) => {
   )
 }
 
-const Crate = (props: BoxProps) => {
-  const [ref, api] = useBox(() => ({ args: [2, 1, 1], mass: 1, ...props }))
+function Crate(props: BoxProps): JSX.Element {
+  const [ref, api] = useBox(() => ({ args: [2, 1, 1], mass: 1, ...props }), useRef<Mesh>(null))
 
   return (
     <Box
@@ -34,33 +34,35 @@ const Crate = (props: BoxProps) => {
   )
 }
 
-const Scene = ({ isPaused = false }) => (
-  <>
-    <OrbitControls />
+function Scene({ isPaused = false }): JSX.Element {
+  return (
+    <>
+      <OrbitControls />
 
-    <Physics gravity={[0, -10, 0]} isPaused={isPaused}>
-      <Debug color="black" scale={1}>
-        <Ground color="grey" position={[0, -2, 0]} rotation={[-Math.PI / 2, 0, 0]} />
+      <Physics gravity={[0, -10, 0]} isPaused={isPaused}>
+        <Debug color="black" scale={1}>
+          <Ground color="grey" position={[0, -2, 0]} rotation={[-Math.PI / 2, 0, 0]} />
 
-        <Crate position={[2, 0, 0]} />
-        <Crate position={[0, 0, 0]} />
-        <Crate position={[-2, 0, 0]} />
+          <Crate position={[2, 0, 0]} />
+          <Crate position={[0, 0, 0]} />
+          <Crate position={[-2, 0, 0]} />
 
-        <Crate position={[2, 5, 0]} />
-        <Crate position={[0, 5, 0]} />
-        <Crate position={[-2, 5, 0]} />
+          <Crate position={[2, 5, 0]} />
+          <Crate position={[0, 5, 0]} />
+          <Crate position={[-2, 5, 0]} />
 
-        <Crate position={[2, 10, 0]} />
-        <Crate position={[0, 10, 0]} />
-        <Crate position={[-2, 10, 0]} />
-      </Debug>
-    </Physics>
+          <Crate position={[2, 10, 0]} />
+          <Crate position={[0, 10, 0]} />
+          <Crate position={[-2, 10, 0]} />
+        </Debug>
+      </Physics>
 
-    <ambientLight />
-  </>
-)
+      <ambientLight />
+    </>
+  )
+}
 
-export default () => {
+export default function Paused() {
   const [isPaused, togglePaused] = useState(false)
   return (
     <div style={{ height: '100vh', width: '100%' }}>